feat: schedule mobile version
