+++ src/components/Schedule/ScheduleMain.js
@@ -7,18 +7,6 @@ import Navbar from "../Navigator/Navigator";
 
 function ScheduleMain() {
   const [schedule, setSchedule] = useState(initialState.schedule);
-  const [innerWidth, setInnerWidth] = useState(window.innerWidth);
-  const [isMobile, setiIsMobile] = useState(false);
-
-  useEffect(() => {
-    const resizeListener = () => {
-      setInnerWidth(window.innerWidth);
-    };
-    window.addEventListener("resize", resizeListener);
-    if (innerWidth < 510) setiIsMobile(true);
-    else setiIsMobile(false);
-    console.log(innerWidth);
-  }, [innerWidth, isMobile]);
 
   useEffect(() => {
     axios
